protein_bo_conformal/requirements.txt
numpy
torch
transformers
fair-esm
